+++ src/main/java/tfw/immutable/ila/longila/LongIlaRamp.java
@@ -38,7 +38,7 @@ public final class LongIlaRamp {
             // }
 
             // INCORRECT, BUT FAST
-            long value = (long) (startValue + increment * start);
+            long value = (startValue + increment * start);
             for (int startInt = (int) start;
                     startInt != startPlusLength;
                     ++startInt, offset += stride, value += increment) {
